feat(#5): refactor about vilna section, create images content and main content components
